+++ R/gs_info_rd.R
@@ -92,6 +92,27 @@
 #' )
 #'
 #' # Example 5 ----
+#' # stratified case under minimal risk weighting and H0: rd0 = 0
+#' gs_info_rd(
+#'   p_c = tibble::tibble(
+#'     stratum = c("S1", "S2", "S3"),
+#'     rate = c(.15, .2, .25)
+#'   ),
+#'   p_e = tibble::tibble(
+#'     stratum = c("S1", "S2", "S3"),
+#'     rate = c(.1, .16, .19)
+#'   ),
+#'   n = tibble::tibble(
+#'     stratum = rep(c("S1", "S2", "S3"), each = 3),
+#'     analysis = rep(1:3, 3),
+#'     n = c(50, 100, 200, 40, 80, 160, 60, 120, 240)
+#'   ),
+#'   rd0 = 0,
+#'   ratio = 1,
+#'   weight = "mr"
+#' )
+#'
+#' # Example 6 ----
 #' # stratified case under sample size weighting and H0: rd0 != 0
 #' gs_info_rd(
 #'   p_c = tibble::tibble(
@@ -109,33 +130,13 @@
 #'   ),
 #'   rd0 = 0.02,
 #'   ratio = 1,
+#'   # users can switch to either "invar" or "mr" weighting as well
 #'   weight = "ss"
-#' )
-#'
-#' # Example 6 ----
-#' # stratified case under inverse variance weighting and H0: rd0 != 0
-#' gs_info_rd(
-#'   p_c = tibble::tibble(
-#'     stratum = c("S1", "S2", "S3"),
-#'     rate = c(.15, .2, .25)
-#'   ),
-#'   p_e = tibble::tibble(
-#'     stratum = c("S1", "S2", "S3"),
-#'     rate = c(.1, .16, .19)
-#'   ),
-#'   n = tibble::tibble(
-#'     stratum = rep(c("S1", "S2", "S3"), each = 3),
-#'     analysis = rep(1:3, 3),
-#'     n = c(50, 100, 200, 40, 80, 160, 60, 120, 240)
-#'   ),
-#'   rd0 = 0.02,
-#'   ratio = 1,
-#'   weight = "invar"
 #' )
 #'
 #' # Example 7 ----
 #' # stratified case under inverse variance weighting and H0: rd0 != 0 and
-#' # rd0 difference for different statum
+#' # rd0 difference for different stratum
 #' gs_info_rd(
 #'   p_c = tibble::tibble(
 #'     stratum = c("S1", "S2", "S3"),
@@ -173,7 +174,7 @@ gs_info_rd <- function(
     ),
     rd0 = 0,
     ratio = 1,
-    weight = c("unstratified", "ss", "invar")) {
+    weight = c("unstratified", "ss", "invar", "mr")) {
   n_analysis <- max(n$analysis)
   weight <- match.arg(weight)
 
@@ -236,6 +237,24 @@ gs_info_rd <- function(
         ) |>
         mutate(weight_per_k_per_s = 1 / sigma2_H1_per_k_per_s / sum_inv_var_per_s) |>
         select(-sum_inv_var_per_s)
+    )
+  } else if (weight == "mr") {
+    suppressMessages(
+      tbl <- tbl |>
+        left_join(
+          tbl |>
+            group_by(analysis) |>
+            summarize(sum_inv_var_per_s = sum(1 / sigma2_H1_per_k_per_s))
+        ) |>
+        ungroup() |>
+        group_by(analysis) |>
+        mutate(alpha_per_k_per_s = (p_c - p_e) * sum_inv_var_per_s - sum((p_c - p_e) / sigma2_H1_per_k_per_s),
+               beta_per_k_per_s = 1/sigma2_H1_per_k_per_s * (1 + alpha_per_k_per_s * sum((p_c - p_e) * n / max(n))),
+               weight_per_k_per_s = beta_per_k_per_s / sum_inv_var_per_s -
+                                       alpha_per_k_per_s / sigma2_H1_per_k_per_s / (sum_inv_var_per_s + sum(alpha_per_k_per_s * (p_c - p_e) / sigma2_H1_per_k_per_s)) *
+                                       sum((p_c - p_e) * beta_per_k_per_s) / sum_inv_var_per_s
+        ) |>
+        select(-c(sum_inv_var_per_s, alpha_per_k_per_s, beta_per_k_per_s))
     )
   }