+++ tests/testthat/test-developer-gs_design_rd.R
@@ -28,3 +28,39 @@ test_that("fixed design", {
 
   expect_equal(x1, x2$analysis$n)
 })
+
+test_that("Stratified GSD: if RD is constant across strata, then MR weights are equal to the INVAR weights", {
+  # Reference: Section 3 of Mehrotra, Devan V., and Radha Railkar.
+  # "Minimum risk weights for comparing treatments in stratified binomial trials." Statistics in Medicine 19.6 (2000): 811-825.
+  x_invar <- gs_design_rd(
+    p_c = tibble::tibble(stratum = c("biomarker positive", "biomarker negative"),
+                         rate = c(.2, .25)),
+    p_e = tibble::tibble(stratum = c("biomarker positive", "biomarker negative"),
+                         rate = c(.15, .20)),
+    rd0 = 0, info_frac = c(0.7, 1),
+    alpha = .025, beta = .1, ratio = 1,
+    stratum_prev = tibble::tibble(stratum = c("biomarker positive", "biomarker negative"),
+                                  prevalence = c(.4, .6)),
+    weight = "invar",
+    upper = gs_spending_bound, lower = gs_b,
+    upar = list(sf = gsDesign::sfLDOF, total_spend = 0.025, param = NULL, timing = NULL),
+    lpar = rep(-Inf, 2))
+
+  x_mr <- gs_design_rd(
+    p_c = tibble::tibble(stratum = c("biomarker positive", "biomarker negative"),
+                         rate = c(.2, .25)),
+    p_e = tibble::tibble(stratum = c("biomarker positive", "biomarker negative"),
+                         rate = c(.15, .20)),
+    rd0 = 0, info_frac = c(0.7, 1),
+    alpha = .025, beta = .1, ratio = 1,
+    stratum_prev = tibble::tibble(stratum = c("biomarker positive", "biomarker negative"),
+                                  prevalence = c(.4, .6)),
+    weight = "mr",
+    upper = gs_spending_bound, lower = gs_b,
+    upar = list(sf = gsDesign::sfLDOF, total_spend = 0.025, param = NULL, timing = NULL),
+    lpar = rep(-Inf, 2))
+
+  expect_equal(x_invar$analysis, x_mr$analysis)
+  expect_equal(x_invar$bound, x_mr$bound)
+})
+